refactor: move InvalidHttpRequestBodyException to teammates.ui.exception
Closes #14075

InvalidHttpRequestBodyException represents an exception, not a request DTO.
Moving it from teammates.ui.request to teammates.ui.exception to match the
project's package organization.

- Move class file to src/main/java/teammates/ui/exception/
- Update package declaration to teammates.ui.exception
- Update all import statements across main, test, and it source trees (46 files)
