+++ src/main/java/teammates/ui/exception/InvalidHttpRequestBodyException.java
@@ -1,4 +1,4 @@
-package teammates.ui.request;
+package teammates.ui.exception;
 
 import teammates.common.exception.InvalidParametersException;
