+++ src/main/java/teammates/ui/servlets/WebApiServlet.java
@@ -11,7 +11,7 @@ import teammates.common.util.HibernateUtil;
 import teammates.common.util.Logger;
 import teammates.ui.exception.InvalidOperationException;
 import teammates.ui.exception.UnauthorizedAccessException;
-import teammates.ui.request.InvalidHttpRequestBodyException;
+import teammates.ui.exception.InvalidHttpRequestBodyException;
 import teammates.ui.webapi.Action;
 import teammates.ui.webapi.ActionFactory;
 import teammates.ui.webapi.ActionResult;
