+++ src/main/java/teammates/ui/webapi/CreateNotificationAction.java
@@ -9,7 +9,7 @@ import teammates.storage.entity.Notification;
 import teammates.ui.exception.InvalidOperationException;
 import teammates.ui.exception.UnexpectedServerException;
 import teammates.ui.output.NotificationData;
-import teammates.ui.request.InvalidHttpRequestBodyException;
+import teammates.ui.exception.InvalidHttpRequestBodyException;
 import teammates.ui.request.NotificationCreateRequest;
 
 /**
