+++ src/main/java/teammates/ui/webapi/PutDataBundleAction.java
@@ -6,7 +6,7 @@ import teammates.common.util.Config;
 import teammates.common.util.JsonUtils;
 import teammates.ui.exception.InvalidOperationException;
 import teammates.ui.exception.UnauthorizedAccessException;
-import teammates.ui.request.InvalidHttpRequestBodyException;
+import teammates.ui.exception.InvalidHttpRequestBodyException;
 
 /**
  * Persists a data bundle into the DB.
