+++ src/main/java/teammates/ui/webapi/SubmitFeedbackResponsesAction.java
@@ -19,7 +19,7 @@ import teammates.ui.exception.UnauthorizedAccessException;
 import teammates.ui.output.FeedbackQuestionResponsesData;
 import teammates.ui.request.FeedbackResponsesRequest;
 import teammates.ui.request.Intent;
-import teammates.ui.request.InvalidHttpRequestBodyException;
+import teammates.ui.exception.InvalidHttpRequestBodyException;
 
 /**
  * Submits feedback responses for one or more feedback questions in a feedback session.
